+++ src/cli/commands/public-cmds/untag-cmd.js
@@ -3,12 +3,13 @@ import Command from '../../command';
 import { unTagAction } from '../../../api/consumer';
 import type { untagResult } from '../../../scope/component-ops/untag-component';
 import GeneralError from '../../../error/general-error';
+import { BASE_DOCS_DOMAIN } from '../../../constants';
 
 const chalk = require('chalk');
 
 export default class Untag extends Command {
   name = 'untag [id] [version]';
-  description = 'revert version(s) tagged for component(s)\n  https://docs.bitsrc.io/docs/cli-untag.html';
+  description = `revert version(s) tagged for component(s)\n  https://${BASE_DOCS_DOMAIN}/docs/cli-untag.html`;
   alias = '';
   opts = [
     ['a', 'all', 'revert tag for all tagged components'],